requirements.txt
begins==0.9
Cartopy==0.14.2
matplotlib==1.5.3
multidict==2.1.2
peru==1.0.1
PyICU==1.9.3
seaborn==0.7.1
